Restructured the Basic example as per convention of other examples.
