+++ examples/Basic/run.py
@@ -1,0 +1,3 @@
+from basic.server import launch_basic
+
+launch_basic()
